Fix the enum validator typing (#3)
Passing an argument of type object into lodash.values results in a
never[] being returned, due to the object type not having keys. With
this fix, only objects/enums with string values can be passed in.

The type of the interface and the validate function differ, because
the Joi.validate function doesn't allow us to add a generic type.